requirements.txt
# requirements.txt
numpy>=1.18
torch>=1.10
torchvision>=0.11
opencv-python>=4.5
Pillow>=8.0
argparse>=1.4
